Adding BaseCode in junit package, code formatting.
