--- src/main/java/junit/App.java
@@ -1,13 +1,0 @@
-package junit;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
