+++ src/test/java/testng/BaseCode.java
@@ -4,9 +4,10 @@ import com.epam.tat.module4.Calculator;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.BeforeMethod;
+
 import java.util.logging.Logger;
 
-public class BaseTest {
+public class BaseCode {
 
     public Calculator calculator;
     Logger logger = Logger.getGlobal();
